Important bugfix for mean phred score calculation
The way the 2-D arrays were constructed was by giving every cell a value of 0, if a cell was 0 after inputting all phred scores then it was counted to not have a character. Because a position can have a score of 0 these are then excluded, thus the calculation of the mean would be incorrect.

Instead of using zeros, np.nan is now used to indicate an empty position.