Insane Assignment1 performance improvements
The `combine_numpy_arrays` has been superseded by the `process_numpy_arrays` function, which now also creates the sum (and count) array. This change was made to allow for the use of integer data types and a new way of creating the count array, that result in massive memory reduction and improved speeds. On a specific configuration it was observed to go from taking 1m11s with a 65GB peak memory usage, to only taking 41s and using 11GB memory.